Use findable midcut for address history hashes
