+++ apps/explorer/src/routes/_layout/address/$address.tsx
@@ -27,6 +27,7 @@ import { BreadcrumbsSlot } from '#comps/Breadcrumbs'
 import { ContractTabContent, InteractTabContent } from '#comps/Contract'
 import { Tip20TokenTabContent } from '#comps/Tip20ContractInfo'
 import { DataGrid } from '#comps/DataGrid'
+import { FindableMidcut } from '#comps/FindableMidcut'
 import { Pagination } from '#comps/Pagination'
 import { Midcut } from 'midcut'
 import { NotFound } from '#comps/NotFound'
@@ -1417,7 +1418,7 @@ function SectionsWrapper(props: {
 												transaction={transaction}
 												accountAddress={address}
 											/>,
-											<Midcut
+											<FindableMidcut
 												key="hash"
 												value={transaction.hash}
 												prefix="0x"
